Fix edit delay not working because of incorrect data type casting.

+++ src/gui/input/SEButton.java
@@ -69,7 +69,7 @@ public class SEButton extends JButton{
     }
 
     private void editDelay() {
-        int delay = (int) SEGUIHandler.instance.delayspin.getValue();
+        double delay = (double) SEGUIHandler.instance.delayspin.getValue();
         int[] indices = SEGUIHandler.instance.list.getSelectedIndices();
         for(int index: indices)
             SEGUIHandler.instance.elm.get(index).setDelay(delay);
